// src/types/cliente.ts
export interface Cliente {
  id: number;
  razaoSocial: string;
  nomeFantasia: string | null;
  cnpj: string;
  porte: string;
  naturezaJuridica: string;
  regime: string;
  inscricaoEstadual: string | null;
  isMei: boolean;
  inscricaoMunicipal: string | null;
  descricaoCNAE: string;
  cnaePrincipal: string;
  prefeitura: string | null;
  anexo1: boolean;
  anexo2: boolean;
  anexo3: boolean;
  anexo4: boolean;
  anexo5: boolean;
  isComercio: boolean;
  isServico: boolean;
  isIndustria: boolean;
  fone: string;
  email: string;
  contatoTel: string | null;
  contatoNome: string | null;
  contatoEmail: string | null;
  representante: string;
  cpfRepresentante: string;
  logradouro: string;
  numero: string;
  complemento: string;
  bairro: string;
  cidade: string;
  estado: string;
  cep: string;
  loginNF: string | null;
  senhaNF: string | null;
  codigoAcessoSN: string | null;
  linkCertificado: string | null;
  senhaCertificado: string | null;
  validadeCertificado: string | null;
  dataCadastro: string | null;
  observacoes: string | null;
  isAtivo: boolean;
}